# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "KAI API",
        "version": "1.0.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "schemes": [
        "http"
    ],
    "paths": {
        "/api/stylist/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "register new stylist",
                "description": "register new stylist",
                "parameters": [
                    {
                        "name": "username",
                        "in": "path",
                        "description": "username",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "email",
                        "in": "path",
                        "description": "email",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "path",
                        "description": "password",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "401": {
                        "description": "validation error"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "User": {
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "nickanme": {
                    "type": "string"
                },
                "age": {
                    "type": "integer"
                },
                "phone_number": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "introduction": {
                    "type": "string"
                },
                "photo_name": {
                    "type": "string"
                }
            }
        }
    }
}